Added spawning of creatures
Added a variable in GameManager that is meant to store the sprites to be
added on the next frame.
Any sprite (for example, an arrow or a knight) that has to be spawned
should be added to lklSpritesToAdd as an instance of the class it is
meant to represent. Then, on update(), a spawning method should be used
to spawn whatever number of instances should be spawned.

Added a public variable called aniDefaultAnim to ResourceManager.java,
which stores a one frame animation. This one can be used as a temporary
substitute for the animation any instance of the class Sprite must have.

Archers now shoot using the forementioned method.

Moved the variable bFacingRight and its respective methods to
Creature.java in order for other creatures to use this variable to
attack correctly.