Added Stickyness to Creature (and enemy shoot frecquency reduced)
Arrows will now stick to any creature they hit.

Every creature can be sticky. This implies that this creature will have
a reference to a specific sprite (the one who it will be stuck to). This
referenced sprite can be used to obtain its position and replace the
creature accordingly to it.